chore(exam): 试卷分数精度迁移 + 黑板进度事件增强 + 各 agent 微调
- Flyway V20 调整 exam 分数相关字段精度（decimal 精度对齐评分粒度）
- BlackboardProgressEvent 增补字段与 builder 方法，支持更细粒度事件
- BlackboardState 新增考点去重报告字段供审计落库
- 六类 agent（Researcher/Scoring/Writer/AnswerKeyGenerator/Reviewer/Calibrator/Deduplicator）
  统一接入黑板事件与调用点，保持接口向后兼容
- ExamHistory / DO / DTO 增加分数字段兼容迁移
- ArticleGenerationSupport 微调
- common.css 管理端通用样式补齐
- FlywayMigrationSmokeTest 同步 V19→V21 迁移计数

+++ knowledge-application/src/main/java/com/mouhin/knowledge/repository/application/executor/articlegeneration/ArticleGenerationSupport.java
@@ -244,7 +244,7 @@ public class ArticleGenerationSupport {
             history.setQuestion(question);
             history.setFinalArticle(blackboard.getFinalArticle());
             history.setDraftArticle(blackboard.getDraftArticle());
-            history.setQualityScore(blackboard.getQualityScore());
+            history.setQualityScore((int) blackboard.getQualityScore());
             history.setRetrievedChunks(retrievedChunks);
             history.setKeyFindings(blackboard.getKeyFindings());
             history.setReviewFeedback(blackboard.getReviewFeedback());
